Poll temperature on a 10s basis
Rather than polling the temperature sensor when the host requests it
use a timer to poll every 10s and store the result. This means we can
reply quicker when requests, reducing the number of errors we see on
the host due to USB issues. Fixes interaction with the double request
python-temperusb does for every check.